pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Agent assistant de planification de voyage - Système multi-agents Google ADK"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["travel_planner"]
